[frontend] Rename components
